Add Media Serving to Backend
Add general media serving to the Haskell server. At URLs beginning with
`/static/`, the server will strip the static prefix, use the remainder
as the path of the file in the Media Directory, & serve the file if it
exists. The Media Directory can be specified by the MEDIA environmental
variable & defaults to a `media` folder in the current working
directory.