[#1108] Add Navbar Categories API Route
This route returns an incomplete set of Categories, the root categories
under a `rootCategories` key, and their children under a
`childrenCategories` key. The children object is a map from a root
category's ID to a list of it's children.

The frontend uses this route to pull the data needed to render the
navbar.

Refs #1108