[#1104] Add Category Details Page
Add a page showing the details of a category, including the name,
description, sub-categories, and sub-products. For now, only products
that are direct children of the category are shown - no products of
child categories are included.

Make Category names on Product Details Pages into links pointing to the
Category's Details Page.

Fix ordering & sizing of middle content.

Refactor the global model to store page-specific data in the Route
instead of as a separate field.

Refactor the view function into multiple functions.

Refs #1104