[#1104] Add Pagination to Category's Products List
Add pagination to the list of Products on a Category's Details Page. The
page number is set & parsed from the URL's query string. The items per
page is also parsed from the query string, but there are currently no
controls on the page to change the value.

Modify the Model & Route types so that fetched data is stored separately
from the route. This allows us to retain data when switching page
numbers(since the route is re-parsed when calling `newUrl`).

Refs #1104