[#1105] Add Quick Search to Header & Search Results Page
Add a search form to the Site Header, allowing customers to search
Product Titles & Descriptions. The results are paginated, but the data
is fetched with a single API query, so the load time is high.

Page speed might be improved by using Esqueleto to reduce the backend
code to a single query. If not, we will have to implement server side
Pagination/Sorting, the client code for this could be re-used from the
FIC Directory pagination.

We will also want to implement Per Page controls & an Advanced search
form. We can special-case the Search Results page for the All Products
page & Organic/Heirloom/Eco/Regional pages.

* Refactor the Product table rendering code into a separate function,
  re-used for the Search Results page.

Refs #1105