[#1130] Add Server-Side Product Pagination & Sorting
Paginate the ProductData in the CategoryDetails & ProductsSearch routes,
defaulting to a page size of 25. This can be overridden by the `perPage`
query parameter(although the frontend does not have controls to modify
this yet) & the page number can by selected by the `page` query
parameter.

The routes now return a `total` field, indicating the total Products in
the non-paginated query.

Sorting mimics the frontend, defaulting to sorting by ascending Product
name. Sorting by price will be implemented when we integrate Esqueleto
into the server.

Refs #1130