[#1105] Add Filters to Server Search API Route
Add the advanced search filters to the Server's Search Route, allowing
filtering by Organic/Heirloom/Regional/Ecological/Category and searching
in only Product Titles. Filtering by Category applies to Products in
sub-categories of the category.

Add the `getChildCategoryIds` helper function to retrieve the IDs of all
it's ancestor categories. This will be useful for showing all the
Products of a Category on the Category Details page instead of only
direct child Products for the Category.

Modify the `paginatedSelect` function's filtering function to take a
potential SeedAttribute as well.

Refs #1105