[#1133] Add All/Organic/Heirloom/SouthEast/Eco Pages
Add the specific product type pages as subsets of the SearchResults
route. Add a UniqueSearch type & uniqueSearch function to the Search
module to detect when a search results page is one of these pages,
modifying the title & hiding the search description.

Add the Icon Legend to the sidebar with links to these pages, as well as
a link to the All Products page in the sidebar navigation.

Refs #1133