[#1149] Return Additional Data for Predecessor Categories
Modify the server's Category Details & Product Details Route so that
thye return the Names & Slugs of their Predecessor Categories as well as
the IDs. This information will be used by the server to generate
breadcrumbs for the Category & Product Details pages.

Replace the getParentCategoryIds function in the Models.Utils module
with getParentCategories, which returns an Entity instead of just the
Category IDs.

Refs #1149