[#1149] Add Breadcrumbs Between Navigation & Site Content
The breadcrumbs are sparse for most pages, but will show a trail of
parent Categories on the Category & Product Details pages.

Modify the PageData module to decode & store the additional Predecessor
Category information returned by the server's Category & Product Details
API routes. Use this data in the SiteUI.Breadcrumbs module to generate
the breadcrumbs. The list of Predecessors returned by the server are
reversed, since they are returned in ascending order instead of the
descending order expected in breadcrumbs.

Closes #1149