[#1109] Add User Registration API Route
Add a route of customer registration, and a route to pull all available
Countries & Regions(for contact addresses). Upon registration, the
customer's password is hashed via high-iteration bcrypt & a
cryptographically-secure v4 UUID is generated as an authorization token.
The parameters to the registration route are the first to validate their
data using the Validation typeclass.

The location route returns all the valid Country/States codes & names.

Refs #1109