[#1109] Add API Route to Authenticate Customers
Add a login route to the server, for authorizing existing Customers. If
authorization is successful, the route will re-hash the password if the
stored hash has a lower cost value than the current high-iteration
standard.

Refactor the RegistrationData type into a AuthorizationData type that is
returned by both the Login & Register routes.

Refs #1109