[#1109] Add Ability To Stay Logged In & To Log Out
Add a "Stay Logged In" checkbox to the Login page, as well as a Log Out
link to the Header. When checked, the User ID & Authentication Token
will be stored in the Browser's Local Storage. When initializing the
site, the saved data will be passed to Elm as flags & will be used to
reauthenticate the Customer.

When the Log Out link is clicked, any saved data will be purged from
Local Storage.

When a tab is used to Log In or Log Out, other tabs on the site will
also log out/in.

Closes #1109