Add Token Authentication Combinator to Server Routing
Add an Auth module to the server, containing the definition of an
`AuthProtect "auth-token"` route combinator which retrieves an AuthToken
from the `Auth-Token` HTTP header or fails with a 401 error.

The module also contains a `validateToken` function for retrieving the
Customer associated with an AuthToken or failing with a 403 error.