Refactor Api Requests & Validation Errors
Refactor the put/post requests into multiple builder functions in the
Api module that construct a Request command. This can be furthur
expanded to define a type for API Endpoints & a function internal to the
tApi module that converts an Endpoint into a URL.

Refactor the Login & Create Account form errors so that validation
errors are pulled out of the Failure case & into the Success case when
the request is sent using the Api.withErrorHandler function.