[#1113] Add Shopping Cart Database Models & Add to Cart API Routes
Add a Cart DB model for organizing an authorized or anonymous customer's
shopping cart. Anonymous carts are identified by a unique token & expire
16 weeks after creation. Add a CartItem DB model for tracking the
ProductVariants & Quantities of items in a Cart.

Add two API routes for adding items to a Cart - one for authorized
customers & one for anonymous customers. Each route will create a Cart
if one does not exist for a registered customer, or if a session token
is not passed to the anonymous route. The anonymous route will always
return the token for the cart that the item was added to.

Refs #1113