[#1169] Add "Cart" Quick Link to Site Header
Add the "Cart (#)" quick link to the site header links, shown only when
the Customer's Cart has items in it. The `#` represents the total
number of items in the cart.

The Cart Item count is stored & loaded from Browser storage, allowing
counts to be shared across tabs. The count is appropriately increased or
decreased when Customers log in, add items to their cart, or
update/remove items from their cart. The count is preserved when a
Customer registers. The count is cleared when a Customer logs out, or
when they are given a new Cart Session Token(because their current token
expired).

The Shopping Cart Details data will be refreshed if on the Shopping Cart
page and a different tab changes the item count.

Factor out some common functions in the javascript entry file.

Closes #1169
Closes #1170