[#1185] Send Order Confirmation Emails
When a successful payment has been made, send the Customer an Order
Confirmation email containing a small message & a summary of their Order
details.

Add the Emails.OrderPlaced module for the new email type. This module
generates an HTML email instead of a Markdown email so we can easily
generate the tables for the Order summary. The Emails module converts
the HTML email into a Markdown plaintext email using Pandoc.

Closes #1185