Fix Creation of Billing Addresses When Same As Shipping
Fix a bug causing no Billing Address to be created during Checkout when
a Customer specifies that their billing Address is the same as their
shipping Address. This required Customers to re-check the box everytime
they checked out or to explicitly enter the Shipping details in the New
Address form.