Fix the Order Totals on the My Account Page
The totals were off because we were only deducting the store credit &
not the member discount. This has been fixed by including all
LineItemTypes that represent credits.

We've added a `creditLineItemTypes` value to the `Models.Fields` module
so that this information can exist in a single place instead of being
littered throughout the code.