Fix Order Totals for Orders with No Credit Lines
Order's with no credit lines were returning totals of $0.00 on the My
Account page. This is because we were using the sum_ function, which
returns Nothing when there are no rows in the result or if all the
values are Null.

When there were no credit line items for an order, this Nothing was
propagating out to the grand total calculation - causing it to return
Nothing, which was then defaulted to 0 cents.

We solve this by adding a wrapper function `sum0_` that defaults the sum
to zero when the result is Nothing and use this for the subTotal,
lineTotal, & creditTotal summations.