[#1223] Remove Contact Fields from Registration Page
The backend does not store the contact information on a per-customer
basis, only in a per-order basis for the billing & shipping information.
However, we were still showing fields for the contact information on the
frontend's registration page.

This fixes the issue by removing these fields, as well as not querying
the backend for the location information.

Fixes #1223