[#1405] Fix Handling of Stripe Errors During Checkout
Fix an issue where Stripe misconfiguration or API errors would cause
orders to be created but not charged. This was caused by intermingling
calls to `runDB` with the Stripe API requests. All order-related
database queries now run in a single database transaction with the
Stripe requests happenining within that transaction. This means that any
Stripe or database errors will cause the transaction to rollback & no
order will be persisted unless all processing steps are completed.

Closes #1405