Do Line Item Calculation in Integers to Prevent Underflow
Modify the createOrder & createLineItems functions in the
Routes.Checkout module so that the line item totals are represented as
Integers.

This is required because the credits may outweigh the charges(e.g., for
free shipping & a member discount) causing the returned charge total to
be negative, which caused an arithmetic underflow exception since Cents
is represented as a Natural(which does not allow negative values).