Add Tests for Coupon Discount Calculations
Add some tests for calculating the Coupon discount for each CouponType.
This requires Show instances for the CartCharge & Coupon types, as well
as exposing the Routes.CommonData module & it's calculateCouponDiscount
function.

This also adds a basic test framework using tasty & hedgehog. The
`build` & `watch` commands on the management script run tests as well.