Add Coupon Code Field Handling to Checkout Page
Add a couponCode field to the Checkout.Form type, as well as Messages
for updating the input coupon code, applying it, and removing an
existing coupon.

Modify the refreshDetails function so that updates can be forced instead
of relying on Form differences.

Submit the couponCode field to the Server when some value has been
entered into the field.

Use couponDiscount returned from server when calculating the checkout
totals.