Add Coupon Code Field to the Checkout Page
Modify the order summary table on the Checkout Page, rendering a coupon
code input to the left of the Order Total. Add an Apply button that
attempts to apply the coupon to the order, rendering any errors that are
returned. If applcation is successful, an additional line item is shown
for the discount applied by the Coupon.