[#1189] Calculate Priority S&H in Checkout Details Route
Add priority shipping & handling calculations to the cusotemr &
anonymous checkout details server routes.

Add a Bool parameter to the routes for indicating whether or not
priority shipping is desired for the Order. If it is desired but not
available, a PriorityShippingNotAvailable error will be thrown.

Refactor the ccShippingMethods field of the CartCharges type to use a
new type, a ShippingCharge. This type stores both the CartCharge &
an optional PriorityShippingFee for a ShippingMethod. The fee is not
present if any of the items are in categories excluded by the
shippingMethodExcludedPriorityCategoryIds field of the ShippingMethod.

Add a ccPriorityShippingFee field to the CartCharges type, containing
the calculated fee for prirority S&H if it was requested.

Modify the getShippingMethods function to return a list of
ShippingCharge instead of a list of CartCharge.

Add HUnit for tests of the new calculatePriorityFee functon.

Refs #1189