[#1189] Add Priority S&H Checkbox to Checkout Page
Add the Priority Shipping & Handling section to the Checkout Page's
form. This contains some basic text about Priority S&H as well as a
checkbox to enable customer's to add it to their order. When the input
is toggled, the details route will be hit to re-calculate the charges &
total and any Priority fee will be shown in the order summary.

Add a priorityShipping field to the CartCharges type to represent it's
line charge. Add PriorityFee & ShippingCharge types to represent the new
shipping method format & the cost of the priority shipping.

Submit the value of the priority checkbox state when requesting the
Checkout Details from the server.

Refs #1189