[#1189] Handle Priority S&H When Placing an Order
Modify the server's Place Order routes so that priority shipping &
handling can be specified in the parameters.

Modify the createLineItems function so it creates a line item for the
PriorityShippingLine if applicable.

Add a Priority Shipping line item to the OrderPlaced email.

Refactor the PriorityShippingNotAvailable error value into a separate
type so it can be used for both the Checkout Details & Place Order
routes.

Refs #1189