[#1189] Show Priority S&H Line Items on Order Details Page
Submit the state of the Priority S&H checkbox when submitting the
Checkout form.

Add a PriorityShipping value to the LineItemType in the PageData along
with a decoder. This allows us to show the Priority Shipping & Handling
line item that is sent by the server for the Order Details page.

Closes #1189