[#1210] Require Less Information for Free Checkouts
Modify the server's checkout routes so that they do not require a Stripe
token or a Billing address if the customer's grand total is equal to
$0. If the total is greater than 0 & these are not provided, an error is
thrown. Refactor the routes so that all the Stripe interactions only
happen if the order total is greater than 0.

Modify the Order Details route, allowing the returned billing address
fields to be null.

Allow the Order database model to have a null Billing Address ID field.

Add a `grandTotal` field to the JSON for the CartCharges type to allow
the client to determine if an Order has a total of $0. Note that any
store credit the customer will apply is not represented here.

Modify the OrderPlaced email so it only attempts to render the Billing
Address if one exists.

Refs #1210