[#1210] Simplify Checkout Page For Free Orders
When an order has a total amount due of 0 dollars(either due to free
products or discounts), hide the Billing Address form, change the submit
button text to "Place Order", & do not collect credit card information
via Stripe.

Modify the placeOrder function so it submits null values for the
billingAddress & stripeToken fields when this is the case.

Add a fullWidth field to the AddressFormConfig type, indicating the
address form should be rendering at the full width of the page, instead
of half-widths like when both the Shipping & Billing forms are present.

Show only the Shipping Address on the Order Details page of free orders.

Decode & utilize the new grandTotal field in the CartCharges type. Note
that we must modify this value according to how much store credit the
customer has applied in the form, since that information is not
submitted to the server until the order is placed.

Closes #1210