[#1398] Add Database Models for Sales
Add CategorySale & ProductSale database models for storing data about
sales. These include the items they apply to, along with a date range in
which they are active. Product sales simply change the price of a
variant, while Category sales can be either a percentage or flat
discount.

Modify the DataMigration script so that old sales are imported from the
ZenCart database.

Refs #1398