[#1398] Apply Sales to Variant Data Returned By Server
Modify the Product, Category, Cart/Checkout routes so instead of
directly returning a ProductVariant from the database, they transform
the ProductVariant into a VariantData value, which contains an extra
field for a potential Sale Price for the variant.

Add functions to the CommonData module for querying the database for
sales and apply any applicable Product & Category sales to the variants.

Refs #1398