[#1398] Add Tests for Category & Product Sales
Add tests to the server for verifying the calculations for both Category
& Product sales. This requires adding a Show instance for the
ProductVariant database model, pulling the DB->API transformation
out into it's own function(makeVariantData), adding new Generator
functions, & exposing the functions in the CommonData module that are
being tested.

Modify the ProductVariant to VariantData transformation so that no sale
price is set if the PercentSale's price is greater than or equal to the
price of the ProductVariant.

Refs #1398