[#1398] Show Sale Prices for Products
Add the salePrice field to the ProductVariant type & add a variantPrice
function that returns the salePrice and falls back to the normal price
when the Variant is not on sale.

Show sale prices in red, below a crossed-out regular price, on the
Category & Product pages. Simply replace the normal price with the sale
price on the Cart, Checkout, & Order Details pages.

Closes #1398