Add Explicit Show/Read Instances for the Cents Type
Adding explicit instances that piggyback off of the Natural type's
instances simplifies the database values for types that use the Cent's
type. While a column with the Cents type was simply an integer field,
any types that had the Cents type nested in them would use the Show
instance when rendering the type for the database, causing more painful
values like `FlatDiscount Cents { fromCents = <value> }` instead of
simply `FlatDiscount <value>`.