[#1439] Properly Render HTML From Model Fields
Replace calls to the Markdown.toHtml function with a new Views.Utils
function, `rawHtml`, which uses the markdown package but disables it's
sanitization process. This allows rendering of the HTML in Product
names, Product/Category descriptions, & StaticPage contents.

Server-side sanitization for these fields will be added to the Admin
pages when they are implemented in a future version.

Closes #1439