[#1445] Fix Page Redraws on Pagination Changes
Fix a bug causing the entire page to re-render when adjusting a
pagination value's sorting or items per page.

The source of the issue was the paginate package's re-initializing of
the Paginated value while changing the requestData or perPage values,
which caused any existing responseData from being removed. This has been
fixed, so we've bumped the version of the dependency.

Modify the view function so that the "Loading..." text for the
CategoryDetails route is only shown if the Category-specific data has
not yet been loaded.

Closes #1445