[#1542] Show Prices for Out of Stock Products
Modify the cartForm for Products so that the price is always shown, even
if the Product is out of stock. To show prices for all variants when
they are out of stock, we show the variant select element and remove the
disabled attributes so the variant & displayed price can be selected.

Closes #1542