[#1159] Add Created Date to Products
Add a createdAt field to the Products DB model and import the created
dates in the DataMigration script. Some imported products have a created
date of 0 - these are shifted to the first day of the Modified Julian
Calendar to support peristent's SQL-to-Haskell conversion function.

Modify the paginatedSelect function to accept sorting values of
"created-asc" & "created-desc".

Refs #1159