[#1128] Add Image Processing Module to Server
Add an Images modules to the server. This module is responsible for
saving, scaling, & optimizing images.

An `ImageConfig` must first be instantiated. This determines which
optimizers are available & whether or not imagemagick is available to
resize images. Currently optipng, jpegtran, & gifsicle is used for
optimization.

The `scaleImage` function takes a filename, image content, & destination
path. The MD5 hash of the content is appended to the filename & the
original image is saved & optimized. Then, the original image is scaled
to the width specified by each `ImageSize` value, saved to a
width-specific folder, & optimized.

Refs #1128