[#1128] Rewrite ImageMigration Script to Scale & Optimize Images
Rewrite the image-migration script so that it uses data from the current
database instead of connecting to the old ZenCart MySQL database. When
migrating images, we hash the file & append the hash to the filename,
scale them to all the sizes specified in the Images module, & optimize
all the resulting images.

Refs #1128