Allow Executing App Monad Actions in SQL Monad
Modify the definition of the AppSQL type so the Transformer is over the
App monad instead of the Handler monad. This allows us to execute
functions that rely on the `Reader Config` type inside SQL transactions,
which will be necessary for serving the scaled media files - since we
check for there existence, which requires knowledge of the Media
Directory.