[#1128] Return Scaled Image Data With Products
Add a new BaseProductData type to hold the Product data required by the
Client, as well as the ImageSourceSet data for a Product generated by
the Images module. Remove the ToJSON & FromJSON instances for the Product
type and replace API usages with the new BaseProductData.

Add a makeBaseProductData to the CommonData module & modify routes that
use to return a Product to use this function to return a BaseProductData
instead.

Refs #1128