[#1403] Support Multiple Lot Size Descriptions
Modify the ProductVariant model so that variants can be described in
other ways than just milligrams by replacing the weight field with a
lotSize field. Add a LotSize type to the Models.Fields module with
built-in support for Mass, Bulb/Slip/Plug counts, custom text, & blank
values.

Modify the routes returning Variants to use the new lotSize field insted
of the weight field.

Modify the DataMigration script so that blank values are given for the
LotSize if the old Product's weight is 0.

Refs #1403