[#1403] Show Lot Sizes Instead of Variant Weights
Modify the client to decode & display the new lotSize field for Product
Variants instead of the removed weight field. A LotSize type has been
added to the Models.Fields module, along with a decoder & string
conversion function.

Modify the milligramsToString function to append the "g" suffix(grams)
to the returned string.

Closes #1403