[#1559] Show Variant LotSizes on Cart & Checkout Pages
Add a nameWithLotSize function to the Product module to nicely render a
Product's name with a ProductVariant's LotSize appended to it if one is
available.

Use this new function to render the product names on the Cart & Checkout
pages, & refactor the product name rendering on the OrderDetails page to
use this as well.

Closes #1559