Improve CI Builds
Modify the management script, adding an isOnTravis Bool field to the
Config type indicating whether the script is being run in CI or not.

If it is run in CI, we install client dependencies using the "npm ci"
command instead of "npm install".

We also indicate that the ~/.npm directory should be cached, which
should improve build times.