[#1541] Show Lot Size in Names of Single-Variant Products
Add a singleVariantName function to the Product module, which appends
the Lot Size to the Product's name if there is only a single variant
present.

Modify the Products.Views module to use this rendering function in the
Product details & list views.

Modify the Product styles to prevent the Lot Size from being underlined
when in hovered links(like the Product name).

Closes #1541