[#1406] Add Login Redirection for Authorized Routes
Improve the flow of authorized pages for anonymous users. When visiting
one of these pages, unauthorized users will be redirected to the Login
page, with a new `redirect` query parameter. After a successful login,
this parameter will be used to redirect user's to the authorized page
they attempted to visit.

Add a parameter to the Route type's Login value for an optional
redirection URL. Parse this from the URL if present. Pass it to the
Auth.Login.view so it can be used to redirect after a successful login.

Modify the redirectIfAuthRequired function so that user's are redirected
to the Login page with the redirect query parameter set to the current
URL.

Closes #1406