[#1150] Improve Error Pages & Fetching Error Display
Modify the View.withIntermediateText function to show better error pages
when the PageData's fetching function returns an error. Split these
pages off into a remoteFailureView function and various helper
functions.

Add a redirect403IfAnonymous function to the Main module, which
redirects Anonymous users to the login page when they get a 403 error
when fetching the data for a page. Modify the Main.update function to
use this helper function when querying data for pages that require a
login.

Add an apiFailureToError function to the Api module for transforming an
Http.Error into a pretty string to include as a Form Error. Use this to
show error messages when form submissions fail due to HTTP errors.

Closes #1150