[#1544] Add Caching for Category Predecessors
Add a Cache module containing types & functions for building a cached
Map from CategoryIds to a list of that Category's predecessor
Categories.

This cache is initialized when the server starts & is stored in a
thread-safe TVar. Resyncing is not necessary at this point, but will be
required when the Category admin page is implemented and the category
hierarchy can change.

Modify the getParentCategories function in the Models.Utils module to
use the CategoryPredecessorCache to get the parent categories instead of
recursive SQL queries.

Add helper functions to the Server module for reading & writing to the
cache.

Closes #1544