Fix Display of Predecessor Categories
Modify the Product Details API route so the Product's parent category is
included in the predecessor list.

Modify the client to properly produce the predecessor lists for the
breadcrumbs & navigation.

This is required because the caching of predecessor categories removed
the current category or product's parent category from the returned
list.