[#1151] Improve Design of Cart Page on Mobile Devices
Revamp the Cart Details page when viewed on mobile screens. Instead of a
cluttered table, we show a block for each CartItem with larger quantity
inputs & update/remove buttons.

We expand the Checkout button at the bottom of the page and remove the
Update button since it is present in each products block.

Refs #1151