[#1151] Redesign Products List Layout on Mobile Screens
Add a separate view for mobile Products Lists in addition to the desktop
one. This shows the products in a single column with horizontal
dividers instead of a table-based layout like the desktop version.

Refactor shared values between the standard cartform & the mobile
cartform into a type alias and generator function, so that the necessary
calculations are only done once when rendering both views. Refactor some
functions for shared functionality between the two views as well.

Refs #1151