[#1142] Add Credentials for StoneEdge Integration to Config
Add a StoneEdgeCredentials type to the StoneEdge module containing
fields for the expected User, Password, & StoreCode in authenticated
routes. Add a HasStoneEdgeCredentials typeclass &
getStoneEdgeRequestCredentials function for pulling the
StoneEdgeCredentials out of a StoneEdge request's data.

Add a field to the Config type for storing the expected credentials &
store code for the StoneEdge integration. Build this field during server
startup.

Refs #1142