[#1399] Migrate API Routes to Cookie Authentication
Change the route declarations in the Customers, Carts, & Checkout route
modules to take a WrappedAuthToken instead of an AuthToken, using either
the withValidatedCookie or validateCookieAndParameters functions to pull
out the AuthToken from the request cookies.

Have the Register, Login, Reset Password, & Anonymous Place Order set a
temporary session cookie when the route processing successful.

Remove the AuthToken from the Authorize route's parameters, pulling it
from the Cookie instead.

In a future commit, the login route will be changed to pick between a
temporary session & permanent session if the user has checked the
"remember me" checkbox. We also need to add a Logout route that removes
the session cookie.

Refs #1399