[#1540] Add Basic Admin UI & Category List Page
Add a client Route for the Categories admin page, as well as the
required data type, a command for fetching the data, and a rendering
function for the page.

Add admin views to the Header & Navigation modules for rendering the
basic UI of the admin site and modify the view function to show the
admin UI if an admin route is loaded.

Refs #1540