[#1540] Add API Route for Creating Categories
Add an API route for Admin's to create new categories. A GET to the
route fetches the names & IDs of existing categories for the Parent
Category dropdown.

Images are passed in with base-64 encoding. The route decodes, saves,
scales, & optimizes them. When the operation is successful, it inserts
the Category into the database & returns the new Category's ID.

Refs #1540