[#1540] Add Admin Page for Creating New Categories
Add an Admin Route & view/model/msg/update for creating new categories.

Move the Admin Category views to a separate module.

Add a textareaRow function to the Views.HorizontalForm module for rendering
textarea elements.

Add a selectImageFile function to the Views.Utils module for generating
commands that prompt the user to select a supported image file.

Refs #1540