[#1540] Add API Route for Editing Categories
Add a editCategoryDataRoute & editCategoryRoute to the Admin API for
fetching a Category's details & updating a Category's fields.

Refactor the makeImage function in the newCategoryRoute into a
makeImageFromBase64 function that is used in both the New & Edit routes.

Modify the NewCategoryParameters validation to ensure the order field is
zero or positive.

Refs #1540