Refactor Pager Rendering into Separate Model
Move the pager rendering code from the Products.Views modules into a
separate Views.Utils module. The new module takes a configuration record
and returns a record of rendering functions for the various sections of
the pager.

This will allow us to re-use the pager rendering for the Admin's Order
List table.