Only Show Admin Pages to Users Authorized as Admins
When a non-admin user tries to visit an Admin page, show them the
standard Access Denied page from the site frontend.

Showing the admin pages was not a security issue since all the API
routes are properly protected, but it did remove the non-admin user from
the normal methods of navigation away from the Access Denied page.