[#1124] Add Order Details Page to the Admin Site
Add an AdminOrderDetails route along with a data fetching & rendering
functions to show Admin users the details of an Order. We re-use the
data types & rendering function for the Customer's OrderDetails route.
However this page will eventually be expanded to include a form allowing
administrators to issue partial/full refunds to Customer.

Add a renderAdminTitle function to conditionally hide the normal page
headers on the admin site. This is used so that we can rely on the
OrderDetails.view function to render the page header.

Refs #1124