Add Order Line Item Type for Refunds
Add a RefundLine value to the LineItemType type for storing refund
amounts. Include these lines in the Order Details page and the Order
Placed email.

Refactor the creditLineItemTypes value to ensure that compilation errors
are thrown when a new LineItemType has been added but the function has
not been updated.