Store Date Alongside Admin Comments
Add a new AdminOrderComment type to the Models.Fields module for storing
both the contents & date/time of an Admin's comment on an Order. This
allows us to be flexible about the date formatting of admin comments in
the frontend & in the StoneEdge integration.

Modify the adminComments field on the Order type to use the new
AdminOrderComment type.

Modify the DataMigration script to pull the comment dates in addition to
the content.

Modify the StoneEdge integration to include date & time prefixes to
admin comments.