[#1124] Add Comment & Refund Forms to Order Details Admin Page
Modify the Views.OrderAdmin module's details function to show any Admin
Comments, as well as forms for adding a new comment & refunding a Stripe
charge. This includes a model, message type, & update function for the
details view.

Add an `id` field to the PageData.Order type.

Add an PageData.AdminOrderDetails type that adds admin comments to the
general OrderDetails type, as well as an AdminOrderComment type which
stores the date & content of admin comments.

Closes #1124