Refactor Where Clause Generation in Orders API
Refactor the building of the SQL `WHERE` clauses into a new
buildWhereQuery function in the Routes.Utils module. This takes the
entire search phrase as well as a function that generates a list of SQL
expressions to OR together for each word in the search phrase.

Move the extractRowCount function from the orderListRoute into the
Routes.Utils function.