[#1123] Add API Routes for Editing Products
Add an editProductDataRoute for fetching the data necessary to show the
Edit Product Form, as well as an editProductRoute for posting the
updated data for a product.

Add ToJSON instances for the NewVariantData & NewSeedData types.

Add an option ProductVariantId field to the NewVariantData type for
updating existing variants of a product.

Extract the duplicate skuSuffix checking function from the
NewProductParameters's Validation instance into a separate function so
it can be re-used in Validation for the EditProductParameters type.

Refs #1123