[#1123] Add Edit Product Page to the Admin Site
Add a new ProductEdit AdminRoute responsible for fetching a Product's
data, rendering & updating a Form, and posting the new data to the API
server.

This admin page behaves differently than previous ones. On the other
pages, the Edit Form keeps track of the fields that have changed with
respect to the original data that was fetched when the page was loaded &
only submits the updated fields. To quicken development, we instead use
the product data as the basis for the form & submit all the product data
to the server instead of only the changed fields. This means the product
data gets stored in the ProductAdmin's EditForm model instead of the
PageData type on the global model, so some tweaking is required in the
Main module's fetchDataForRoute function so that the
model.editProductForm gets updated instead of the model.pageData field.

Add links to the edit page to the Admin's Products List page/table.
Redirect users to the product's edit page after they successfully create
a new product.

Modify the ProductAdmin.formView function so that the submit button's
text is passed in as an argument. Show an existing image in the form if
one is specified.

Closes #1123