[#1532] Try ZenCart Passwords for Initial Logins
Implement ZenCart's password hashing policy in the API server's
loginRoute. It is only attempted if the customerEncryptedPassword field
matches the format of ZenCarts passwords & normal hashing does not
validate the password. If the ZenCart password is valid, the stored
password hash is upgraded to the normally used BCrypt policy.

This keeps us from requiring password resets from all migrated
Customers.

Closes #1532